flask_backend/requirements.txt
Flask==2.2.2
werkzeug==2.2.3
mysql-connector-python==8.0.30
